Move grep.rs to grep/mod.rs
