+++ .github/workflows/ci.yml
@@ -58,7 +58,7 @@ jobs:
       - uses: actions-rs/clippy-check@v1
         with:
           token: ${{ secrets.GITHUB_TOKEN }}
-          args: --all-features
+          args: --all-features --manifest-path crates/Cargo.toml
 
   crates:
     name: Rust Tests
